add thor for simplify
