+++ Gemfile
@@ -2,3 +2,4 @@ source 'https://rubygems.org'
 
 # Specify your gem's dependencies in pt.gemspec
 gemspec
+gem 'aruba', '~> 0.14.2'
