+++ bin/pt
@@ -7,4 +7,4 @@ $LOAD_PATH.unshift(lib) if File.directory?(lib) && !$LOAD_PATH.include?(lib)
 
 require 'pt'
 
-PT::UI.new ARGV
+PT::CLI.start ARGV
