+++ lib/pt.rb
@@ -4,7 +4,8 @@ require 'pt/client'
 require 'pt/data_row'
 require 'pt/data_table'
 require 'pt/action'
-require 'pt/ui'
+require 'pt/helper'
+require 'pt/cli'
 
 module PT
    # Your code goes here...
