+++ spec/support/aruba.rb
@@ -1,0 +1,1 @@
+# require 'aruba/rspec'
